Compare FCA Call Report metadata files between two folders

compare_metadata() compares the content of the metadata files (files that start with "D_") between two specified folders containing FCA Call Report data (from two different quarters).

Usage

compare_metadata(dir1, dir2)

Arguments

dir1

(String) The path to a folder containing FCA Call Report .TXT files for a single quarter

dir2

(String) The path to a folder containing FCA Call Report .TXT files for a (different) single quarter

Value

A list containing information about differences in file names, file order, and content differences between the metadata files in dir1 and dir2

Details

compare_metadata() lists metadata files in each folder, identifies shared metadata files, and then compares (a) the number of files, (b) file names, (c) file order, and (d) file content (using the waldo::compare() function).

Examples

# \donttest{

  # Download data from September 2025
  path_1 <- tempfile("fcadata1")
  dir.create(path_1)

  download_data(
    year = 2025,
    month = 9,
    dest = path_1
  )
#> Files successfully downloaded into /tmp/RtmpRzwt5d/fcadata11b025b321f99

  # Download data from September 2011
  path_2 <- tempfile("fcadata2")
  dir.create(path_2)

  download_data(
    year = 2011,
    month = 9,
    dest = path_2
  )
#> Files successfully downloaded into /tmp/RtmpRzwt5d/fcadata21b02422bb2a5

  compare_metadata(path_1, path_2)
#> $file_differences
#>      dir1               | dir2             
#>  [2] "D_RC.TXT"         | "D_RC.TXT"   [2] 
#>  [3] "D_RC1.TXT"        | "D_RC1.TXT"  [3] 
#>  [4] "D_RCB.TXT"        | "D_RCB.TXT"  [4] 
#>  [5] "D_RCB2.TXT"       - "D_RCF.TXT"  [5] 
#>  [6] "D_RCB3.TXT"       - "D_RCF1.TXT" [6] 
#>  [7] "D_RCB4.TXT"       - "D_RCG.TXT"  [7] 
#>  [8] "D_RCB5.TXT"       - "D_RCH.TXT"  [8] 
#>  [9] "D_RCF.TXT"        - "D_RCI.TXT"  [9] 
#> [10] "D_RCF1.TXT"       - "D_RCK.TXT"  [10]
#> [11] "D_RCG.TXT"        - "D_RCL.TXT"  [11]
#> [12] "D_RCH.TXT"        - "D_RCM.TXT"  [12]
#> [13] "D_RCI1_2018.TXT"  - "D_RCO.TXT"  [13]
#> [14] "D_RCI2A_2018.TXT" - "D_RI.TXT"   [14]
#> [15] "D_RCI2B_2018.TXT" - "D_RIA.TXT"  [15]
#> [16] "D_RCI2C_2018.TXT" - "D_RIB.TXT"  [16]
#> [17] "D_RCI2D_2018.TXT" - "D_RIC.TXT"  [17]
#> [18] "D_RCK.TXT"        - "D_RIC1.TXT" [18]
#> [19] "D_RCL.TXT"        - "D_RID.TXT"  [19]
#> [20] "D_RCM.TXT"        - "D_RIE.TXT"  [20]
#> [21] "D_RCO.TXT"        -                  
#> [22] "D_RCR1.TXT"       -                  
#> [23] "D_RCR2.TXT"       -                  
#> [24] "D_RCR3.TXT"       -                  
#> [25] "D_RCR4.TXT"       -                  
#> [26] "D_RCR5.TXT"       -                  
#> [27] "D_RCR6.TXT"       -                  
#> [28] "D_RCR7.TXT"       -                  
#> [29] "D_RI.TXT"         -                  
#> [30] "D_RIA.TXT"        -                  
#> [31] "D_RIB.TXT"        -                  
#> [32] "D_RIC.TXT"        -                  
#> [33] "D_RIC1.TXT"       -                  
#> [34] "D_RID.TXT"        -                  
#> [35] "D_RIE1.TXT"       -                  
#> [36] "D_RIE2.TXT"       -                  
#> 
#> $content_differences
#> $content_differences$D_RC.TXT
#> dir1[2:95] vs dir2[2:85]
#>   "                              DATA DELIMITED BY COMMA"
#>   "                 "
#>   "                               "
#> - "         VARIABLE     FIELD  DEC.               "
#> - "             NAME      TYPE  POS.  VARIABLE DESCRIPTION "
#> - "    --------------  -------  ----  --------------------"
#> - "           SYSTEM   Numeric     0  System Code   "
#> - "             DIST   Numeric     0  District Code"
#> - "            ASSOC   Numeric     0  Association Code"
#> - "            MONTH   Numeric     0  Month of Report"
#> - "             YEAR   Numeric     0  Year of Report"
#> - "           UNINUM   Numeric     0  System, District, and Association"
#> - "                                    codes concatenated"
#> - "             CASH   Numeric     0  Cash"
#> - "           ACTREC   Numeric     0  Accounts Receivable"
#> - "           ACRLNS   Numeric     0  Accrual Loans and Leases net of"
#> - "                                    unearned income and unapplied"
#> - "                                    loan payments"
#> - "        NTRECOFCI   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                    leases: Notes receivable from other"
#> - "                                    FCS institutions"
#> - "           ONTREC   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                  leases: Other notes receivable"
#> - "          ACRSCON   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                    leases: Accrual Sales Contracts"
#> - "           NONACR   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                    leases: Nonaccrual Loans "
#> - "         ALLNLOSS   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                  leases: Allowance for credit losses on loans, etc."
#> - "         NETLOANS   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                    leases: Net Loans "
#> - "         AIRLOANS   Numeric     0  Accrued Interest Receivable on Loans"
#> - "                                    and Leases  "
#> - "     AIRNTRECOFCI   Numeric     0  Accrued Interest Receivable on Notes"
#> - "                                    receivable from other FCS Institutions"
#> - "        AIRONTREC   Numeric     0  Accrued Interest Receivable on Other"
#> - "                                    Notes receivable"
#> - "          AIRSCON   Numeric     0  Accrued Interest Receivable on Sales"
#> - "                                    Contracts"
#> - "         AIRMKTIN   Numeric     0  Accrued Interest Receivable on"
#> - "                                    Securities"
#> - "            AIREC   Numeric     0  Total Accrued Interest Receivable"
#> - "          NAQPROP   Numeric     0  Other Property Owned "
#> - "         FIXASSET   Numeric     0  Premises and Other Fixed Assets"
#> - "          OASSETS   Numeric     0  Other Assets "
#> - "           ASSETS   Numeric     0  Total Assets   "
#> - "       SYSNTBDSOS   Numeric     0  Interest bearing liabilities:"
#> - "                                    Systemwide notes and bonds outstanding"
#> - "        NTPAYOFCI   Numeric     0  Interest bearing liabilities:"
#> - "                                    Notes payable to other FCS"
#> - "       OINTBEARDT   Numeric     0  Interest bearing liabilities:"
#> - "                                    Other interest-bearing debt "
#> - "           ACTPAY   Numeric     0  Accounts Payable"
#> - "    AIPSYSNTBDSOS   Numeric     0  Accrued interest payable on: Systemwide"
#> - "                                    notes and bonds outstanding"
#> - "     AIPNTPAYOFCI   Numeric     0  Accrued interest payable on: Notes"
#> - "                                    payable to other FCS institutions"
#> - "    AIPOINTBEARDT   Numeric     0  Accrued Interest Payable on Other"
#> - "                                    interest-bearing debt      "
#> - "            AIPAY   Numeric     0  Total Accrued Interest Payable "
#> - "           OLIABS   Numeric     0  Other Liabilities    "
#> - "            LIABS   Numeric     0  Total Liabilities "
#> - "          NETWRTH   Numeric     0  Total Net Worth"
#> - "         TOTLICAP   Numeric     0  Total Liabilities and Net Worth"
#> - "       ACNOCOMINC   Numeric     0  Accumulated other comprehensive income (net)"
#> - "        LNADJFVAL   Numeric     0  Loans, notes, sales, contracts, and"
#> - "                                    leases: Loan adjustment for fair value"
#> - "        DTADJFVAL   Numeric     0  Interest bearing liabilities:"
#> - "                                    Debt adjustment for fair value"
#> - "   TOTINTBEARLIAB   Numeric     0  Interest bearing liabilities:"
#> - "                                    Total interest bearing liabilities "
#> - "    AIRDERIVATIVE   Numeric     0  Accrued interest receivable: Derivatives"
#> - "   IBLSUBNOTEBOND   Numeric     0  Interest bearing liabilities: "
#> - "                                    Subordinated notes and bonds"
#> - "   AIPSUBNOTEBOND   Numeric     0  Accrued interest payable on: "
#> - "                                    Subordinated notes and bonds"
#> - "    AIPDERIVATIVE   Numeric     0  Accrued interest payable on: Derivatives"
#> - "    LoanLeaseHeld   Numeric     0  Loans and leases held for sale"
#> - "CreditLossReserve   Numeric     0  Reserve for credit losses on off-balance"
#> - "\t\t\t\t\t\t\t\t\tsheet exposures"
#> - "      TERMPERFSTK   Numeric     0  Interest bearing liabilities:"
#> - "\t\t\t\t\t\t\t\t\tLimited-life (term) preferred stock"
#> - "AccrdIntPayTermPrefStk Numeric  0  Accrued interest payable on:"
#> - "\t\t\t\t\t\t\t\t\tLimited-life (term) preferred stock"
#> - "\t\t CapStock   Numeric     0  Capital stock"
#> - " PerpetualPrefStk   Numeric     0  Perpetual Preferred stock"
#> - "\t\t   PAIDIN  \tNumeric     0  Paid-in capital"
#> - "\t AllocSurplus   Numeric     0  Allocated Surplus"
#> - "\t\t UNRETERN   Numeric     0  Unallocated retained earnings"
#> - "TotInvDebtSecurities Numeric    0  Debt securities, net of allowance for credit losses "
#> - "TotInvEquityFCSI \t Numeric    0  Equity investments in System institutions "
#> - "\t\t\t\t\t\t\t\t\tand Farmer Mac"
#> - "TotInvEquityNonFCSI\t Numeric    0  Equity investments in non-System institutions"
#> - "TotInvestments\t\t Numeric    0  Total Investments"
#> + "       VARIABLE    FIELD  DEC.               "
#> + "           NAME     TYPE  POS.  VARIABLE DESCRIPTION "
#> + "  -------------- -------  ----  --------------------"
#> + "         SYSTEM  Numeric     0  System Code   "
#> + "           DIST  Numeric     0  District Code"
#> + "          ASSOC  Numeric     0  Association Code"
#> + "          MONTH  Numeric     0  Month of Report"
#> + "           YEAR  Numeric     0  Year of Report"
#> + "         UNINUM  Numeric     0  System, District, and Association"
#> + "                                 codes concatenated"
#> + "           CASH  Numeric     0  Cash"
#> + "         MKTINV  Numeric     0  Securities"
#> + "         ACTREC  Numeric     0  Accounts Receivable"
#> + "         ACRLNS  Numeric     0  Accrual Loans and Leases net of"
#> + "                                 unearned income and unapplied"
#> + "                                 loan payments"
#> + "      NTRECOFCI  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Notes receivable from other"
#> + "                                 FCS institutions"
#> + "         ONTREC  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Other notes receivable"
#> + "        ACRSCON  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Accrual Sales Contracts"
#> + "         NONACR  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Nonaccrual Loans "
#> + "       ALLNLOSS  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Allowance for Losses on Loans"
#> + "       NETLOANS  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Net Loans "
#> + "       AIRLOANS  Numeric     0  Accrued Interest Receivable on Loans"
#> + "                                 and Leases  "
#> + "   AIRNTRECOFCI  Numeric     0  Accrued Interest Receivable on Notes"
#> + "                                 receivable from other FCS Institutions"
#> + "      AIRONTREC  Numeric     0  Accrued Interest Receivable on Other"
#> + "                                 Notes receivable"
#> + "        AIRSCON  Numeric     0  Accrued Interest Receivable on Sales"
#> + "                                 Contracts"
#> + "       AIRMKTIN  Numeric     0  Accrued Interest Receivable on"
#> + "                                 Securities"
#> + "          AIREC  Numeric     0  Total Accrued Interest Receivable"
#> + "       INFCINST  Numeric     0  Equity Investments in other FC"
#> + "                                 Institutions"
#> + "        NAQPROP  Numeric     0  Other Property Owned "
#> + "       FIXASSET  Numeric     0  Premises and Other Fixed Assets"
#> + "        OASSETS  Numeric     0  Other Assets "
#> + "         ASSETS  Numeric     0  Total Assets   "
#> + "     SYSNTBDSOS  Numeric     0  Interest bearing liabilities:"
#> + "                                 Systemwide notes and bonds outstanding"
#> + "      NTPAYOFCI  Numeric     0  Interest bearing liabilities:"
#> + "                                 Notes payable to other FCS"
#> + "     OINTBEARDT  Numeric     0  Interest bearing liabilities:"
#> + "                                 Other interest-bearing debt "
#> + "         ACTPAY  Numeric     0  Accounts Payable"
#> + "  AIPSYSNTBDSOS  Numeric     0  Accrued interest payable on: Systemwide"
#> + "                                 notes and bonds outstanding"
#> + "   AIPNTPAYOFCI  Numeric     0  Accrued interest payable on: Notes"
#> + "                                 payable to other FCS institutions"
#> + "  AIPOINTBEARDT  Numeric     0  Accrued Interest Payable on Other"
#> + "                                 interest-bearing debt      "
#> + "          AIPAY  Numeric     0  Total Accrued Interest Payable "
#> + "         OLIABS  Numeric     0  Other Liabilities    "
#> + "          LIABS  Numeric     0  Total Liabilities "
#> + "         CAPSTK  Numeric     0  Capital Stock, Participation"
#> + "                                 certificates and paid-in surplus"
#> + "                                 net of impairment"
#> + "         EARNNW  Numeric     0  Earned Surplus Net of Impairment"
#> + "        NETWRTH  Numeric     0  Total Net Worth"
#> + "       TOTLICAP  Numeric     0  Total Liabilities and Net Worth"
#> + "     ACNOCOMINC  Numeric     0  Accumulated other comprehensive income (net)"
#> + "      LNADJFVAL  Numeric     0  Loans, notes, sales, contracts, and"
#> + "                                 leases: Loan adjustment for fair value"
#> + "      DTADJFVAL  Numeric     0  Interest bearing liabilities:"
#> + "                                 Debt adjustment for fair value"
#> + " TOTINTBEARLIAB  Numeric     0  Interest bearing liabilities:"
#> + "                                 Total interest bearing liabilities "
#> + "  AIRDERIVATIVE  Numeric     0  Accrued interest receivable: Derivatives"
#> + " IBLSUBNOTEBOND  Numeric     0  nterest bearing liabilities: "
#> + "                                 Subordinated notes and bonds"
#> + " AIPSUBNOTEBOND  Numeric     0  Accrued interest payable on: "
#> + "                                 Subordinated notes and bonds"
#> + "  AIPDERIVATIVE  Numeric     0  Accrued interest payable on: Derivatives"
#> 
#> $content_differences$D_RC1.TXT
#> dir1[55:71] vs dir2[55:95]
#>   "  AMTLNBKRUPTCY   Numeric     0  Total amount of loans in bankruptcy"
#>   "AMTLNFORECLOSUR   Numeric     0  Total amount of loans in foreclosure"
#>   "   TAMTBANKFORE   Numeric     0  Total amount of loans in both "
#>   "                                  bankruptcy and foreclosure"
#> + "       INFCIPUR   Numeric     0  Association investment in FCS district "
#> + "                                  bank - purchased"
#> + "     INFCIALLOC   Numeric     0  Association investment in FCS district "
#> + "                                  bank - allocated"
#> + "  FCIEQAMTCPRMC   Numeric     0  Association investment in FCS district "
#> + "                                  bank - Allocation counted as association"
#> + "                                  permanent capital"
#> + "  FCIBKINDASSOC\t  Numeric     0  Bank investment in district "
#> + "                                  associations      "
#> + "      FCIINVOBK   Numeric     0  Investment in other FCBs and ACBs"
#> + "     FCIINVOFCI   Numeric     0  Investment in other Farm Credit "
#> + "                                  institutions"
#> + "  NETASSOCINVBK   Numeric     0  Net association investment in district bank"
#> + "   FCIINVNONFCI   Numeric     0  Equity investments in non Farm Credit "
#> + "                                  institutions"
#>   "       FCIBONDS   Numeric     0  Farm Credit investment bonds "
#>   "                                  outstanding"
#> + "    TERMPERFSTK   Numeric     0  Term preferred stock"
#>   "    INVFARMNOTE   Numeric     0  Investments in farmers' notes"
#>   "     NPDIRLNTBK   Numeric     0  Notes payable/direct loan to district "
#>   "                                  bank"
#>   "   FCSICPREMPAY   Numeric     0\t Farm Credit System Insurance "
#>   "                                  Corporation premium payable"
#> - "\t\t TYPOTH\t  Numeric     0  Other loans"
#> - "       LMPRIN90   Numeric     0  Deprecated\t\t\t\t "
#> - "          LMAMT   Numeric     0  Deprecated"
#> - "           LMNO\t  Numeric     0  Deprecated"
#> + "       PERMCAPA   Numeric     0  Permanent capital amount (outstanding)"
#> + "    CORESURPAMT   Numeric     0  Core surplus amount"
#> + "     TOTSURPAMT   Numeric     0  Total surplus amount"
#> + "       PERMCAPD   Numeric     4  Permanent capital ratio"
#> + "  CORESURPRATIO   Numeric     4  Core surplus ratio"
#> + "   TOTSURPRATIO   Numeric     4  Total surplus ratio"
#> + "   NETCOLLRATIO   Numeric     4  Net collateral ratio"
#> + "\t TYPOTH\t  Numeric     0  Other loans"
#> + "       LMPRIN90   Numeric     0  Principal portion of obligations maturing"
#> + "\t\t\t\t  within 90 days"
#> + "          LMAMT   Numeric     0  Dollar amount of liquidity at quarter-end"
#> + "           LMNO\t  Numeric     0  Number of days liquidity at quarter-end"
#>   "   TYPPROCMKT12\t  Numeric     0  Processing and Marketing Loans:"
#>   "                                  Loans made under 613.3010(a)(1) and (a)(2) "
#>   "    TYPPROCMKT3   Numeric     0  Processing and Marketing Loans: "
#> 
#> dir1[75:91] vs dir2[99:106]
#>   "    TYPPROCMKT5   Numeric     0  Processing and Marketing Loans: "
#>   "                                  Loans made under 613.3010(a)(5)"
#>   "     TOP10LOANS   Numeric     0  Aggregate amount of institution’s                                  "
#> - "\t\t\t\t\t\t\t\t 10 largest loan commitments"
#> - "NUMLOANSFRMFDIC   Numeric  \t  0  Number of loans outstanding pur"
#> - "\t\t\t\t\t\t\t\t\tchased from the FDIC"
#> - "AMTLOANSFRMFDIC   Numeric  0\t Amount of loans outstanding purchased "
#> - "\t\t\t\t\t\t\t\t\tfrom the FDIC"
#> - "    LLLIMITBASE   Numeric     0  Lending and Leasing Limit Base (Stops collecting as of March 2022)"
#> - "LoanLeaseHeldAmortizedCost   Numeric     0  Loans and Leases Held for Sale: Amortized cost   "
#> - "    LoanLeaseHeldFairValue   Numeric     0  Loans and Leases Held for Sale: Fair value"
#> - "\t\t\t\t TotRegCap\t Numeric     0  Lending and Leasing Limit Base:"
#> - "\t\t\t\t\t\t\t\t\t\t\t\tTotal regulatory capital (Sch. RC-R.1, item 5b.)"
#> - "\t\t Excess3rdPartyCap\t Numeric     0  Lending and Leasing Limit Base:"
#> - "\t\t\t\t\t\t\t\t\t\t\t\tEligible excess third-party capital above regulatory capital limits"
#> - "\t LendLeaseLimitBaseTRC\t Numeric     0  Lending and Leasing Limit Base:"
#> - "\t\t\t\t\t\t\t\t\t\t\t\tLending and Leasing Limit Base (sum of items 5a. and 5b.)\t\t\t\t\t\t\t\t\t"
#> + "\t\t\t\t  10 largest loan commitments"
#> + "     NUMLOANSFRMFDIC Numeric  0  Number of loans outstanding pur"
#> + "\t\t\t\t  chased from the FDIC"
#> + "     AMTLOANSFRMFDIC Numeric  0\t Amount of loans outstanding purchased "
#> + "\t\t\t\t  from the FDIC"
#> 
#> $content_differences$D_RCB.TXT
#> dir1[1:6] vs dir2[1:5]
#> - "         FILE LAYOUT FOR SCHEDULE RC-B Debt Securities"
#> - "\t\t\t\t\t\t(excluding investments in FC institutions and"
#> - "\t\t\t\t\t\tdiversified investment funds)  "
#> + "         FILE LAYOUT FOR SCHEDULE RC-B MARKETABLE INVESTMENTS BY TYPE"
#> + "                              DATA DELIMITED BY COMMAS  "
#>   "                                               "
#>   "                                      "
#>   "       VARIABLE    FIELD  DEC.     "
#> 
#> dir1[15:72] vs dir2[14:81]
#>   "                                 codes concatenated"
#>   "   **  INV_CODE  Numeric     0  Investment Code for Marketable"
#>   "                                 Investments: "
#> - "\t\t\t\t10  U.S. Treasury securities"
#> - "                15  SBA securities"
#> - "\t\t\t\t17  Other U.S. Gov’t securities and Agency(excluding MBS)"
#> - "\t\t\t\t"
#> - "                25  Securities fully and unconditionally guaranteed by a"
#> - "\t\t\t\t\tGSE (excluding MBS and Farmer Mac securities)"
#> - "\t\t\t\t29  Municipal securities"
#> - "                35  International and multilateral development"
#> - "\t\t\t\t\tbank obligations"
#> - "\t\t\t\t\t"
#> - "\t\t\t\t40  Money market instruments:Federal funds sold"
#> - "\t\t\t\t41  Money market instruments:Negotiable certificates of"
#> - "\t\t\t\t\t\t\t\t\t\t\tDeposit"
#> - "                60  Money market instruments:Banker acceptances"
#> - "                52  Money market instruments:Commercial paper"
#> - "                                       "
#> - "                50  Money market instruments:Reverse repurchase agreements"
#> - "                51  Money market instruments:Other"
#> - ""
#> - "                62  RMBS fully and unconditionally guaranteed by U.S.  "
#> - "\t\t\t\t\tgovernment or its agencies"
#> - "\t\t\t\t68  RMBS fully and unconditionally guaranteed by GSE"
#> - "\t\t\t\t64  Non-Agency RMBS"
#> - "\t\t\t\t69  Other RMBS"
#> - "\t\t\t\t"
#> - "\t\t\t\t71  CMBS(excl. Farmer Mac sec.):CMBS fully and unconditionally "
#> - "\t\t\t\t\t\t\t\t\t\t\t\tguaranteed by U.S. government"
#> - "                72  CMBS(excl. Farmer Mac sec.):CMBS fully and unconditionally "
#> - "\t\t\t\t\t\t\t\t\t\t\t\tguaranteed by GSE"
#> - "\t\t\t\t73  CMBS(excl. Farmer Mac sec.):Non-Agency CMBS"
#> - "                65  CMBS(excl. Farmer Mac sec.):Other CMBS"
#> - "\t\t\t\t"
#> - "\t\t\t\t86  Farmer Mac guaranteed sec.:Farm and ranch securities (i.e., AMBS)"
#> - "\t\t\t    87  Farmer Mac guaranteed sec.:Rural utility securities"
#> - "                88  Farmer Mac guaranteed sec.:USDA securities"
#> - "                66  Farmer Mac guaranteed sec.:Other Farmer Mac debt securities"
#> - ""
#> - "\t\t\t    141 ABS(excl. Farmer Mac securities):Credit card receivables"
#> - "\t\t\t    142 ABS(excl. Farmer Mac securities):Home equity loans"
#> - "\t\t\t    143 ABS(excl. Farmer Mac securities):Auto Loans"
#> - "\t\t\t    144 ABS(excl. Farmer Mac securities):Student Loans"
#> - "\t\t\t    145 ABS(excl. Farmer Mac securities):Equipment loans"
#> - "\t\t\t    146 ABS(excl. Farmer Mac securities):Manufactured housing loans"
#> - "\t\t\t\t95  ABS(excl. Farmer Mac securities):Other ABS"
#> - "\t\t\t\t"
#> - "\t\t\t\t81  Other types of debt securities:Domestic debt securities"
#> - "\t\t\t\t82  Other types of debt securities:Foreign debt securities"
#> - "\t\t\t\t"
#> - "\t\t\t\t180 Allowance for Credit Losses on Debt Securities(start collecting on March 2023)"
#> - "\t\t\t\t"
#> - "\t\t\t\t99  Total Debt Securities"
#> - "\t\t\t\t   "
#> + "                                   10  U.S. Treasury securities"
#> + "                                   11  Other fully insured or guaranteed"
#> + "                                       U.S. Government obligations"
#> + "\t\t\t\t   12 Other partially insured or guaranteed "
#> + "\t\t\t\t\tobligations"
#> + "                                   20  U.S. Government agency securities"
#> + "\t\t\t\t   22 Other partially insured or guaranteed "
#> + "\t\t\t\t\tobligation"
#> + "                                   31  General Municipal obligations"
#> + "\t\t\t\t   33 Industrial development revenue bonds"
#> + "\t\t\t\t   34 Other"
#> + "                                   32  Municipal Revenue Bonds"
#> + "                                   35  International and multilateral"
#> + "                                       Development Bank obligations"
#> + "                                   40  Federal funds sold"
#> + "                                   41  Negotiable certificates of"
#> + "                                       Deposit"
#> + "                                   43  Non-callable term Federal funds"
#> + "                                       and Euro time deposits"
#> + "                                   50  Securities purchased under"
#> + "                                       resale agreements"
#> + "                                   52  Money market instruments:"
#> + "                                       commercial paper"
#> + "                                   60  Banker acceptances"
#> + "                                   62 Fully insured or guaranteed by "
#> + "\t\t\t\t\tU.S. Government – NEW TITLE                                "
#> + "                                   64  Non-Agency Mortgage securities"
#> + "\t\t\t\t   65  Commercial MBS"
#> + "                                   66 Agricultural MBS issued or "
#> + "\t\t\t\t\tguaranteed by Farmer MAC"
#> + "                                   68 Fully insured or guaranteed by US GSA"
#> + "\t\t\t\t   69 Securities partially insured or guaranteed by US Government or GSA"
#> + "\t\t\t\t   70  Other"
#> + "                                   80  Total"
#> + "\t\t\t\t   81 Domestic debt securities"
#> + "\t\t\t           82 Foreign debt securities"
#> + "                                   85  Diversified investment funds"
#> + "                                   90  Master Notes"
#> + "                                   95 Other structured securities –NEW TITLE"
#> + "                                   100 Pledged securities"
#> + "                                   141 Credit card receivables"
#> + "\t\t\t\t   142 Home equity loans"
#> + "\t\t\t\t   143 Auto Loans"
#> + "\t\t\t\t   144 Student Loans"
#> + "\t\t\t\t   145 Equipment loans"
#> + "\t\t\t\t   146 Manufactured housing loans"
#> + "\t\t\t\t   150 Securities subject to the 35% "
#> + "\t\t\t\t\tregulatory limit"
#> + "\t\t\t\t   151 Securities included in the "
#> + "\t\t\t\t\tliquidity reserve"
#> + "\t\t\t\t   152 Securities initially purchased "
#> + "\t\t\t\t\tfor the liquidity reserve, "
#> + "\t\t\t\t\tbut are no longer \t\t\t\t\teligible to be included in the liquidity reserve"
#> + "\t\t\t\t   153 Other securities that are not "
#> + "\t\t\t\t\tincluded in the liquidity reserve"
#> + "\t\t\t\t   154 Cash and securities in liquidity "
#> + "\t\t\t\t\treserve (fair value)"
#> + "\t\t\t\t   155 Cash and securities in liquidity "
#> + "\t\t\t\t\treserve, discounted in accordance with  \t\t\t\t\t615.5134"
#> + "\t\t\t\t   156 Principal portion of obligations maturing "
#> + "\t\t\t\t\twithin 90 days"
#> + "\t\t\t\t   157 Total number days liquidity "
#>   "          "
#>   "    **     BKVAL   Numeric    0  Amortized cost of Marketable Investments"
#>   "    **    MKTVAL   Numeric    0  Fair Value of Marketable Investments  "
#> 
#> $content_differences$D_RCF1.TXT
#> dir1[26:33] vs dir2[26:33]
#>   "                                 140 Lease receivables"
#>   "                                 145 Direct loans to associations"
#>   "                                 150 Discounted loans to OFIs"
#> - "\t\t\t\t 152 Other loans"
#> + "\t\t\t\t 155 Other loans"
#> - "                                 155 Total"
#> + "                                 160 Total"
#>   "  **       ACCR  Numeric     0  Accruing Loans"
#>   "  **   ACCRPDUE  Numeric     0  Accruing 90 days or more past due"
#>   "  **    FRMREST  Numeric     0  Formally restructured accruing"
#> 
#> $content_differences$D_RCG.TXT
#> dir1[2:35] vs dir2[2:40]
#>   "                           DATA DELIMITED BY COMMAS"
#>   "                                            "
#>   "                      "
#> - "\t\t\t\t   VARIABLE    FIELD  DEC.     "
#> - "\t\t\t\t\t   NAME     TYPE  POS.  VARIABLE DESCRIPTION                    "
#> - "\t\t\t---------------  -------  ----  ---------------------                   "
#> - "\t\t\t\t\t SYSTEM  Numeric     0  System Code    \t       \t                "
#> - "\t\t\t\t\t   DIST  Numeric     0  District Code                           "
#> - "\t\t\t\t\t  ASSOC  Numeric     0  Association Code                        "
#> - "\t\t\t\t\t  MONTH  Numeric     0  Month of Report               "
#> - "\t\t\t\t\t   YEAR  Numeric     0  Year of Report              "
#> - "\t\t\t\t\t UNINUM  Numeric     0  System, District, and Association   "
#> - "\t\t\t\t\t\t\t\t\t\t\t codes concatenated"
#> - "\t\t\t\t   AVACRLNS  Numeric     0  Accrual loans, sales contracts, notes,"
#> - "\t\t\t\t\t\t\t\t\t\t\t and leases"
#> - "\t\t\t\t   AVAIRLNS  Numeric     0  Accrued interest receivable (loans)"
#> - "\t\t\t AVGOEARNASSETS  Numeric     0  Other Interest-Earning Assets \t   "
#> - "\t\t\t\t   AVEARNAS  Numeric     0  Total Interest-Earning Assets"
#> - "\t\t\t\t   AVNONACB  Numeric     0  Nonaccrual loans - Cash basis"
#> - "\t\t\t\t   AVNONACO  Numeric     0  Nonaccrual loans - Other  "
#> - "\t\t\t\t   AVAQPROP  Numeric     0  Other property owned"
#> - "\t\t\t\t\tAVASSET  Numeric     0  Total Assets         "
#> - "\t\t\t\t AVSYSNTBDS  Numeric     0  Systemwide notes and bonds"
#> - "\t\t\t\tAVNTPAYOFCI  Numeric     0  Notes payable to other FCS institutions"
#> - "\t\t\t  AVGOBEARLIABS  Numeric     0  Other Interest-Bearing Liabilities"
#> - "\t\t\t\t   AVBEARLI  Numeric     0  Total Interest-Bearing Liabilities"
#> - "\t\t\t\t   AVGLIABS  Numeric     0  Total Liabilities    "
#> - "\t\t\t\t\tAVNWRTH  Numeric     0  Total net worth "
#> - "\t\t\tAvLoanLeaseHeld  Numeric     0  Loans and leases held for sale"
#> - "  AvgDailyInvDebtSecurities\t Numeric     0  Investments: Debt securities"
#> - " \t  AvgDailyInvEquityFCSI  Numeric     0  Investments: Equity investments in "
#> - "\t\t\t\t\t\t\t\t\t\t\tSystem institutions and Farmer Mac"
#> - "   AvgDailyInvEquityNonFCSI\t Numeric     0  Investments: Equity investments in "
#> - "\t\t\t\t\t\t\t\t\t\t\tnon-System institutions"
#> + "       VARIABLE    FIELD  DEC.     "
#> + "           NAME     TYPE  POS.  VARIABLE DESCRIPTION                    "
#> + "---------------  -------  ----  ---------------------                   "
#> + "         SYSTEM  Numeric     0  System Code    \t       \t                "
#> + "           DIST  Numeric     0  District Code                           "
#> + "          ASSOC  Numeric     0  Association Code                        "
#> + "          MONTH  Numeric     0  Month of Report               "
#> + "           YEAR  Numeric     0  Year of Report              "
#> + "         UNINUM  Numeric     0  System, District, and Association   "
#> + "                                 codes concatenated"
#> + "       AVACRLNS  Numeric     0  Accrual loans, sales contracts, notes,"
#> + "                                 and leases"
#> + "       AVAIRLNS  Numeric     0  Accrued interest receivable (loans)"
#> + "       AVMKTINV  Numeric     0  Securities"
#> + " AVGOEARNASSETS  Numeric     0  Other Interest-Earning Assets \t   "
#> + "       AVEARNAS  Numeric     0  Total Interest-Earning Assets"
#> + "       AVNONACB  Numeric     0  Nonaccrual loans - Cash basis"
#> + "       AVNONACO  Numeric     0  Nonaccrual loans - Other  "
#> + "          AVGFR  Numeric     0  Formally Restructured Loans "
#> + "     AVDLNASSOC  Numeric     0  Direct loans to associations"
#> + "    AVNTRECOFCI  Numeric     0  Notes receivable from other FCS"
#> + "                                 institutions"
#> + "       AVAQPROP  Numeric     0  Other property owned"
#> + "        AVRSKAS  Numeric     0  Risk-adjusted assets"
#> + "        AVASSET  Numeric     0  Total Assets         "
#> + "     AVSYSNTBDS  Numeric     0  Systemwide notes and bonds"
#> + "    AVNTPAYOFCI  Numeric     0  Notes payable to other FCS institutions"
#> + "  AVGOBEARLIABS  Numeric     0  Other Interest-Bearing Liabilities"
#> + "       AVBEARLI  Numeric     0  Total Interest-Bearing Liabilities"
#> + "       AVGLIABS  Numeric     0  Total Liabilities    "
#> + "      AVDIRLNBK  Numeric     0  Notes payable/direct loan to district"
#> + "                                 bank"
#> + "       AVPRMCAP  Numeric     0  Permanent capital"
#> + "        AVNWRTH  Numeric     0  Total net worth "
#> + "       AVCORSUR  Numeric     0  Core surplus"
#> + "       AVTOTSUR  Numeric     0  Total surplus  "
#> 
#> $content_differences$D_RCH.TXT
#> dir1[11:21] vs dir2[11:43]
#>   "          MONTH   Numeric     0  Month of Report"
#>   "           YEAR   Numeric     0  Year of Report"
#>   "         UNINUM   Numeric     0  System, District, and Association   "
#> - "                                  codes concatenated             "
#> - "     ACNOCOMINC   Numeric     0  Total accumulated other comprehensive"
#> - "     UNREALGAIN   Numeric     0\t Net unrealized gains (or losses) on securities"
#> - "\t\t\t\t\t\t\t\t   available-for-sale \t\t\t\t"
#> - "      OTHIMPSEC   Numeric     0\t Non-credit related impairment on securities available-for-sale"
#> - "     PENSIONADJ   Numeric     0\t Minimum pension liability adjustments"
#> - "   CASHHEDGEADJ   Numeric     0\t Cash flow hedge adjustments"
#> - "      OTHINCADJ   Numeric     0\t Other comprehensive income adjustments"
#> + "                                  codes concatenated"
#> + "           STKP   Numeric     0  Capital Stock - protected"
#> + "           STKO   Numeric     0  Capital Stock - other                "
#> + "      PARTCERTP   Numeric     0  Participation Certificates - protected"
#> + "      PARTCERTO   Numeric     0  Participation Certificates - other"
#> + "     PREFSTKFAC   Numeric     0  Preferred Stock-FCSFAC"
#> + "       PREFSTKO   Numeric     0  Preferred Stock-Other"
#> + "         PAIDIN   Numeric     0  Paid-In Capital, etc."
#> + "         CAPSTK   Numeric     0  Total Capital Stock"
#> + "        ALSURPP   Numeric     0  Allocated Surplus-Protected"
#> + "        ALSURPO   Numeric     0  Allocated Surplus-Other"
#> + "         EARNNW   Numeric     0  Total Earned Surplus"
#> + "        NETWRTH   Numeric     0  Total Net Worth"
#> + "     ACNOCOMINC   Numeric     0  Total accumulated other comprehensive"
#> + "      CAPSTKPUR   Numeric     0  Capital Stock - Purchased"
#> + "    CAPSTKALLOC   Numeric     0  Capital Stock - Allocated"
#> + "    PREFSTKPERP   Numeric     0  Perpetual"
#> + "     ALLOCEQOTH   Numeric     0  Allocated equities allotted to others"
#> + "     ALLOCEQNCS   Numeric     0  Allocated equities included in core surplus"
#> + "       UNRETERN   Numeric     0  Unallocated retained earnings"
#> + "       \t  NQWNA   Numeric     0  Nonqualified written notice of allocation - "
#> + "        NQALSUR\t  Numeric     0  Earned Surplus: Nonqualified Allocated Surplus\t\t\t  "
#> + "\t\t\t\t\tunallocated"
#> + "     UNREALGAIN   Numeric     0\t Net unrealized losses (or gains) on securities \t\t\t\tavailable-for-sale that are not other-than \t\t\t\ttemporarily impaired"
#> + "      OTHIMPSEC   Numeric     0\t Other-than-temporarily impaired available-for-\t\t\t\tsale securities"
#> + "     PENSIONADJ   Numeric     0\t Minimum pension liability adjustments"
#> + "   CASHHEDGEADJ   Numeric     0\t Cash flow hedge adjustments"
#> + "      OTHINCADJ   Numeric     0\t Other comprehensive income adjustments"
#> + "\t\t\t\t     \tIncome\t"
#> + ""
#> 
#> $content_differences$D_RCO.TXT
#> dir1[16:22] vs dir2[16:21]
#>   "                                   10  Loan participations: Purchased"
#>   "                                   20  Loan participations: Sold"
#>   "                                   30  Similar entity transactions: Acquired"
#> - "\t\t\t\t       On Interest Held"
#>   "                                   40  Similar entity transactions: Sold"
#>   "                                   50  Lease interest purchases and sales:"
#>   "                                       Purchased"
#> 
#> dir1[30:39] vs dir2[29:35]
#>   "                                       Purchased"
#>   "\t\t\t\t  100  Participations in Notes Receivables:"
#>   "                                       Sold\t"
#> - "\t\t\t\t  110  Asset Purchases and Sales - Certain Pool Items:"
#> - "\t\t\t\t       Purchased"
#> - "\t\t\t\t  120  Asset Purchases and Sales - Certain Pool Items:"
#> - "                                       Sold"
#> + ""
#>   "             "
#>   "**    TRANSWFCI  Numeric     0  Amortized cost"
#>   "** TRANSWNONFCI  Numeric     0  Fair Value"
#> 
#> $content_differences$D_RI.TXT
#> dir1[29:36] vs dir2[29:36]
#>   "                                 bearing debt       "
#>   "      TOTINTEXP  Numeric     0  Total Interest Expense"
#>   "        NETINTY  Numeric     0  Net Interest Income"
#> - "        PROVLNS  Numeric     0  Provisions for credit losses:"
#> + "        PROVLNS  Numeric     0  Provisions for Losses on Loans,Sales"
#> - "\t\t\t\t\t\t\t\t\tOn loans, sales contracts, notes, and leases"
#> + "                                 Contracts,Notes,and Leases"
#>   "        PROVOTH  Numeric     0  Provisions for Losses on Other Assets"
#>   "         PATDIV  Numeric     0  Patronage,Dividends,etc.,from other"
#>   "                                 FC Institutions"
#> 
#> dir1[63:73] vs dir2[64:67]
#>   "                                 (or loss(-))"
#>   "        COMPINC  Numeric     0  Comprehensive income (or loss(-))"
#>   " CASHFLOHEDGADJ  Numeric     0  Cash flow hedge adjustments"
#> - "  NONTEMPIMPAIR\t numeric     0  Non-credit impairment on securities available-for-sale\t"
#> - "  \t\t\t\tincluded in earnings"
#> - "        PATPAID\t Numeric     0  Patronage on Asset Purchases and Sales - Certain Pool Programs:"
#> - "\t\t\t        Paid"
#> - "    PATRECEIVED  Numeric     0  Patronage on Asset Purchases and Sales - Certain Pool Programs:"
#> - "\t\t\t\tRecieved"
#> - "LossProvisionOBSExposure  Numeric     0  Provisions for losses: On off-balance sheet exposures"
#> - "\tProvDebtSec  Numeric\t0\tProvisions for credit losses: On debt securities (start collecting on March 2023)"
#> + " NONTEMPIMPAIRN\t numeric     0  Other-than-temporary impairments\t"
#> 
#> $content_differences$D_RIB.TXT
#> dir1[30:38] vs dir2[30:33]
#>   "        GAINOTRAN  Numeric     0  Gains on Other Transactions"
#>   "        LOSSOTRAN  Numeric     0  Losses on Other Transactions"
#>   "      NETGLASSETS  Numeric     0  Total Gains or Losses"
#> - "    NETIMPAIRLOSS  Numeric     0  Impairment on available-for-sale debt securities "
#> - "LoanLeaseSaleNetGain  Numeric     0  Net gains (or losses(-)) on sales of "
#> - "\t\t\t\t      loans and leases"
#> - "   LoanLeaseSaleGain  Numeric     0  Gains"
#> - "   LoanLeaseSaleLoss  Numeric     0  Losses"
#> - "   "
#> + "    NETIMPAIRLOSs  Numeric     0  Net other-than-temporary impairment losses \t"
#> 
#> $content_differences$D_RID.TXT
#> dir1 vs dir2
#> - "\t                 FILE LAYOUT FOR SCHEDULE RI-D CHANGES IN NET WORTH"
#> + "                 FILE LAYOUT FOR SCHEDULE RI-D CHANGES IN NET WORTH"
#> - "\t                             DATA DELIMITED BY COMMAS"
#> + "                             DATA DELIMITED BY COMMAS"
#>   "    "
#>   "                                "
#> - "\t         VARIABLE    FIELD  DEC.         "
#> + "         VARIABLE    FIELD  DEC.         "
#> - "\t             NAME     TYPE  POS.  VARIABLE DESCRIPTION"
#> + "             NAME     TYPE  POS.  VARIABLE DESCRIPTION"
#> - "\t  ---------------  -------  ----  --------------------------------------"
#> + "  ---------------  -------  ----  --------------------------------------"
#> - " \t          SYSTEM  Numeric     0  System Code"
#> + "           SYSTEM  Numeric     0  System Code"
#> - "\t             DIST  Numeric     0  District Code"
#> + "             DIST  Numeric     0  District Code"
#> - "\t            ASSOC  Numeric     0  Association Code"
#> + "            ASSOC  Numeric     0  Association Code"
#> - "\t            MONTH  Numeric     0  Month of Report"
#> + "            MONTH  Numeric     0  Month of Report"
#> - "\t             YEAR  Numeric     0  Year of Report"
#> + "             YEAR  Numeric     0  Year of Report"
#> - "\t           UNINUM  Numeric     0  System, District, and Association "
#> + "           UNINUM  Numeric     0  System, District, and Association "
#>   "                                   codes concatenated"
#> - "\t     **  CAP_CODE  Numeric     0  Capital Stock Code: "
#> - "\t                                   10  Beginning balance"
#> - "\t                                   25  Prior Period & Accounting Adjustments"
#> - "\t                                   35  Net Income"
#> - "\t                                   45  Other Comprehensive Income"
#> - "       \t                                   80  Patronage Distributions"
#> - " \t   \t                           70  Dividends"
#> - "  \t                                   75  Stock Issued"
#> - "                                \t   85  Stock Retired"
#> - "                        \t           95  Paid-in Capital Adjustments"
#> - "                \t                   105 Allocated Equity Retired"
#> - "        \t                           120 Other"
#> - "\t                                   130 Ending balance"
#> - "    **        CapStkPurch  Numeric     0  Capital Stock:"
#> - "\t\t\t\t\t    Purchased"
#> - "    **        CAPSTKALLOC  Numeric     0  Capital Stock:"
#> - "\t\t\t\t\t    Allocated"
#> - "    **        PREFSTKPERP  Numeric     0  Perpetual Preferred Stock"
#> - "\t\t\t\t\t    Noncumulative"
#> - "    **           PREFSTKO  Numeric     0  Perpetual Preferred Stock"
#> - "\t\t\t\t\t    Other"
#> - "    **    \t   PAIDIN  Numeric     0  Paid-In Capital"
#> - "    **     \t    ALSUR  Numeric     0  Allocated Surplus:"
#> - "\t\t\t\t\t    Qualified"
#> - "    **    NQASSubj2Retire  Numeric     0  Allocated Surplus:"
#> - "\t\t\t\t\t    Nonqualified Subject To Retirement"
#> - "    ** NQASNotSubj2Retire  Numeric     0  Allocated Surplus:"
#> - "\t\t\t\t\t    Nonqualified Not Subject To Retirement"
#> - "    **           UNRETERN  Numeric     0  Unallocated Retained Earnings"
#> - "    **         ACNOCOMINC  Numeric     0  Accumulated other comprehensive income (AOCI)"
#> - "    **\t            TOTNW  Numeric     0  Total Net Worth"
#> - ""
#> + "     **  CAP_CODE  Numeric     0  Capital Stock Code: "
#> + "                                   10  Beginning balance"
#> + "                                   20  Prior quarter adjustments"
#> + "                                   30  Amended beginning balance"
#> + "                                   40  Accum. adjustments acctg. change"
#> + "                                   50  Prior year's adj."
#> + "                                   60  Net income"
#> + "                                   70  Dividents"
#> + "                                   80  Patronage dist."
#> + "                                   90  Other income transfers"
#> + "                                   100 Retirements of cap. stock, part."
#> + "                                       cert. and surplus allocated"
#> + "                                   110 Issuance of cap. stk. and part."
#> + "                                       cert."
#> + "                                   115 Issued pref. stk."
#> + "                                   117 Retired pref. stk."
#> + "                                   120 Other"
#> + "                                   130 Ending balance"
#> + "     **       CAP  Numeric     0  Capital Stock and Participation"
#> + "                                   Certificates"
#> + "     **    PAIDIN  Numeric     0  Paid-In Capital"
#> + "     **     ALSUR  Numeric     0  Surplus Allocated"
#> + "     **     SURES  Numeric     0  Surplus Reserve"
#> + "     **    UALSUR  Numeric     0  Earned Surplus Unallocated"
#> + "     **     TOTNW  Numeric     0  Total Net Worth"
#> + "     **  NONALSUR  Numeric     0  Nonqualified allocated surplus"
#> + " ** OTHCOMPINCOME  Numeric     0  Accumulated Other Comprehensive Income"
#> + "  "
#>   "                                                   "
#>   "  ** NOTE:  THE RECORD CONTAINS MULTIPLE OCCURANCES OF THESE"
#>   "            VARIABLES.  THERE IS ONE OCCURENCE FOR EACH CAP_CODE VALUE."
#> 
#> 

# }